수의 비밀 구현

+++ 19년 9월/9월 1주차/수의 비밀.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int main(){
-    int a;
+    long long a;
     cin>>a;
     while(a>1){
         if(a%2){
